admin-service/app/api/dependencies.py
import jwt
from fastapi import Depends, HTTPException
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials

from app.config.settings import settings
from app.database.repositories.admin_repository import AdminRepository
from app.utils.logger import logger

# Security
security = HTTPBearer()

def verify_jwt_token(token: str) -> dict:
    """Verify JWT token and return payload."""
    try:
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.JWTError:
        raise HTTPException(status_code=401, detail="Invalid token")

async def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(security)):
    """Get current user from JWT token."""
    token = credentials.credentials
    payload = verify_jwt_token(token)
    
    admin_repository = AdminRepository()
    user = admin_repository.get_user_by_id(payload["user_id"])
    
    if not user:
        raise HTTPException(status_code=401, detail="User not found")
    
    return user

async def get_admin_user(current_user: dict = Depends(get_current_user)):
    """Ensure current user has admin privileges."""
    if not current_user.get("is_admin", False):
        raise HTTPException(status_code=403, detail="Admin access required")
    
    logger.info(f"Admin user {current_user.get('email')} accessed admin endpoint")
    return current_user