admin-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pyjwt==2.8.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart==0.0.6